chore: Build out README and add .flox
Adding more details in the README as this project looked interesting, but I had
to poke at it a bit to figure out exactly what it was doing. This hopefully
helps clear that up. Also documents a bit about getting a SLACK_TOKEN.

This also adds a .flox to the setup.